refactor: extract constants into separate files
