+++ packages/twenty-front/src/modules/workflow/workflow-diagram/constants/NodeBorderWidth.ts
@@ -1,0 +1,1 @@
+export const NODE_BORDER_WIDTH = 1;
