+++ packages/twenty-front/src/modules/workflow/workflow-diagram/constants/NodeHandleHeightPx.ts
@@ -1,0 +1,3 @@
+import { NODE_HANDLE_WIDTH_PX } from './NodeHandleWidthPx';
+
+export const NODE_HANDLE_HEIGHT_PX = NODE_HANDLE_WIDTH_PX;
