+++ packages/twenty-front/src/modules/workflow/workflow-diagram/constants/NodeHandleWidthPx.ts
@@ -1,0 +1,1 @@
+export const NODE_HANDLE_WIDTH_PX = 4;
