+++ packages/twenty-front/src/modules/workflow/workflow-diagram/constants/NodeIconLeftMargin.ts
@@ -1,0 +1,5 @@
+import { THEME_COMMON } from 'twenty-ui';
+
+export const NODE_ICON_LEFT_MARGIN = Number(
+  THEME_COMMON.spacing(2).replace('px', ''),
+);
