+++ packages/twenty-front/src/modules/workflow/workflow-diagram/constants/NodeIconWidth.ts
@@ -1,0 +1,5 @@
+import { THEME_COMMON } from 'twenty-ui';
+
+export const NODE_ICON_WIDTH = Number(
+  THEME_COMMON.spacing(6).replace('px', ''),
+);
